Fix missing validation on Job input and other services that got lost in the shuffle
