+++ genie-core/src/main/java/com/netflix/genie/core/services/JobSubmitterService.java
@@ -20,6 +20,7 @@ package com.netflix.genie.core.services;
 
 import com.netflix.genie.common.dto.JobRequest;
 import com.netflix.genie.common.exceptions.GenieException;
+import org.springframework.validation.annotation.Validated;
 
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
@@ -30,6 +31,7 @@ import javax.validation.constraints.NotNull;
  * @author amsharma
  * @since 3.0.0
  */
+@Validated
 public interface  JobSubmitterService {
 
     /**
