+++ genie-web/src/main/java/com/netflix/genie/web/configs/ServicesConfig.java
@@ -49,6 +49,7 @@ import com.netflix.genie.core.services.impl.CacheGenieFileTransferService;
 import com.netflix.genie.core.services.impl.DefaultMailServiceImpl;
 import com.netflix.genie.core.services.impl.FileSystemAttachmentService;
 import com.netflix.genie.core.services.impl.GenieFileTransferService;
+import com.netflix.genie.core.services.impl.JobCoordinatorServiceImpl;
 import com.netflix.genie.core.services.impl.LocalJobKillServiceImpl;
 import com.netflix.genie.core.services.impl.LocalJobRunner;
 import com.netflix.genie.core.services.impl.MailServiceImpl;
@@ -351,7 +352,7 @@ public class ServicesConfig {
         final ApplicationEventPublisher eventPublisher,
         final String hostName
     ) {
-        return new JobCoordinatorService(
+        return new JobCoordinatorServiceImpl(
             taskExecutor,
             jobPersistenceService,
             jobSubmitterService,
